+++ src/main/java/pie/tomato/tomatomarket/infrastructure/persistence/item/ItemPaginationRepository.java
@@ -11,7 +11,7 @@ import com.querydsl.core.types.Projections;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
 import lombok.RequiredArgsConstructor;
-import pie.tomato.tomatomarket.infrastructure.persistence.PaginationRepository;
+import pie.tomato.tomatomarket.infrastructure.persistence.util.PaginationUtil;
 import pie.tomato.tomatomarket.presentation.request.item.ItemResponse;
 
 @RequiredArgsConstructor
@@ -20,7 +20,6 @@ public class ItemPaginationRepository {
 
 	private final JPAQueryFactory queryFactory;
 	private final ItemRepository itemRepository;
-	private final PaginationRepository paginationRepository;
 
 	public Slice<ItemResponse> findByIdAndRegion(Long itemId, String region, int size, Long categoryId) {
 		List<ItemResponse> itemResponses = queryFactory.select(Projections.fields(ItemResponse.class,
@@ -42,6 +41,6 @@ public class ItemPaginationRepository {
 			.orderBy(item.createdAt.desc())
 			.limit(size + 1)
 			.fetch();
-		return paginationRepository.checkLastPage(size, itemResponses);
+		return PaginationUtil.checkLastPage(size, itemResponses);
 	}
 }
